+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## main
 
-### dependency
-- remove dependency to `networkx`
+### backend API
+- add two new parameters to `AudioLabeling` constructor: `audio` and `annotations`. These two parameters can be used to init an `AudioLabeling` component with specified audio and annotations, for visualization purposes for instance.
+```python
+audio = ...       # str, Path or (int, numpy.ndarray)
+annotations = ... # pyannote.core.Annotation
+audio_labeling = AudioLabeling(type="filepath", interactive=True, audio=audio, annotations=annotations)
+```
 
 ### improvement
 - move time cursor so that active segment is always visible on the screen
 - improve `AudioLabeling` zoom interface
+
+### dependency
+- remove dependency to `networkx`